MC-31983: Refactor code to match the agreed repo module structure
- Fixed review comments on tests and repository files
