+++ dev/tests/integration/testsuite/Magento/CustomerMessageBroker/_files/customer.php
