+++ dev/tests/integration/testsuite/Magento/CustomerMessageBroker/_files/customer_address_data.php
