+++ dev/tests/integration/testsuite/Magento/CustomerMessageBroker/_files/customer_with_address.php
