+++ dev/tests/integration/testsuite/Magento/CustomerSynchronizer/_files/customer.php
