+++ dev/tests/integration/testsuite/Magento/CustomerSynchronizer/_files/customer_rollback.php
