+++ dev/tests/integration/testsuite/Magento/CustomerSynchronizer/_files/customer_with_address_rollback.php
