+++ dev/tests/integration/testsuite/Magento/CustomerSynchronizer/_files/customer_with_no_rolling_back.php
